Fixes phpunit and phpstan issues and raises the phpstan level to 10 Signed-off-by: Joey Smith <jsmith@webinertia.net>
