+++ phpstan.neon.dist
@@ -16,4 +16,4 @@ parameters:
         - stubs/Psr/Container/ContainerInterface.stub
     treatPhpDocTypesAsCertain: false
     typeAliases:
-        TracyConfig: 'array{enable?: bool|string|null, productionMode?: bool|null, showBar?: bool, reservedMemorySize?: int, strictMode?: bool|int, scream?: bool|int, maxDepth?: int, maxLength?: int, maxItems?: int, showLocation?: bool|null, keysToHide?: string[], dumpTheme?: string, logDirectory?: string|null, logSeverity?: int, email?: string|string[]|null, editor?: string|null, editorMapping?: array<string, string>, browser?: string|null, errorTemplate?: string|null, customCssFiles?: string[], customJsFiles?: string[]}'
+        TracyConfig: 'array{enable?: bool, productionMode?: bool|null, showBar?: bool, reservedMemorySize?: int, strictMode?: bool|int, scream?: bool|int, maxDepth?: int, maxLength?: int, maxItems?: int, showLocation?: bool|null, keysToHide?: string[], dumpTheme?: string, logDirectory?: string|null, logSeverity?: int, email?: string|string[]|null, editor?: string|null, editorMapping?: array<string, string>, browser?: string|null, errorTemplate?: string|null, customCssFiles?: string[], customJsFiles?: string[]}'
