+++ src/Debug/ConfigPanel.php
@@ -14,15 +14,38 @@ declare(strict_types=1);
 
 namespace Webware\Traccio\Debug;
 
+use Tracy\Helpers;
 use Tracy\IBarPanel;
 
-final class ConfigPanel implements IBarPanel
+final readonly class ConfigPanel implements IBarPanel
 {
-    use IBarPanelTrait;
+    private string $id;
 
+    /**
+     * @param array<mixed> $data
+     */
     public function __construct(
         private array $data,
     ) {
         $this->id = 'config';
+    }
+
+    public function getTab(): string
+    {
+        return Helpers::capture(function () {
+            $data  = $this->data;
+            $title = $this->id;
+
+            require __DIR__ . "/panels/{$this->id}.tab.phtml";
+        });
+    }
+
+    public function getPanel(): string
+    {
+        return Helpers::capture(function () {
+            $data = $this->data;
+
+            require __DIR__ . "/panels/{$this->id}.panel.phtml";
+        });
     }
 }
