+++ src/Debug/RoutesPanel.php
@@ -15,15 +15,35 @@ declare(strict_types=1);
 namespace Webware\Traccio\Debug;
 
 use Mezzio\Router\RouteCollectorInterface;
+use Tracy\Helpers;
 use Tracy\IBarPanel;
 
-final class RoutesPanel implements IBarPanel
+final readonly class RoutesPanel implements IBarPanel
 {
-    use IBarPanelTrait;
+    private string $id;
 
     public function __construct(
         private RouteCollectorInterface $data,
     ) {
         $this->id = 'routes';
+    }
+
+    public function getTab(): string
+    {
+        return Helpers::capture(function () {
+            $data  = $this->data;
+            $title = $this->id;
+
+            require __DIR__ . "/panels/{$this->id}.tab.phtml";
+        });
+    }
+
+    public function getPanel(): string
+    {
+        return Helpers::capture(function () {
+            $data = $this->data;
+
+            require __DIR__ . "/panels/{$this->id}.panel.phtml";
+        });
     }
 }
